// client/packages/web-widgets/src/app/widgets/grid-filters/enums/operator-symbol.ts
export enum OperatorSymbol {
  Contains = 'contains',
  NotContains = 'notContains',
  Equals = 'equals',
  NotEqual = 'notEqual',
  LessThan = 'lessThan',
  LessThanOrEqual = 'lessThanOrEqual',
  GreaterThan = 'greaterThan',
  GreaterThanOrEqual = 'greaterThanOrEqual',
  DateBefore = 'lessThan',
  DateOnBefore = 'lessThanOrEqual',
  DateAfter = 'greaterThan',
  DateOnAfter = 'greaterThanOrEqual'
}